% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{launch_app}
\alias{launch_app}
\title{Launch Interactive Shiny App}
\usage{
launch_app()
}
\description{
Launches an interactive Shiny application to explore the multi-path AIC
selection procedure with visualizations.
}
\examples{
\dontrun{
launch_app()
}
}
